upadte init sql

+++ connect-common/src/main/java/com/connect/common/util/ImageUploadUtil.java
@@ -13,6 +13,7 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.List;
 
 @Slf4j
 @Component
@@ -20,22 +21,42 @@ public class ImageUploadUtil {
     @Value("${imageUploadUtil.profileImage.directory:./static/profileImage}")
     public String profileImageDir;
 
+    @Value("#{'${imageUploadUtil.valid.extension:jpeg,jpg,png,bmp}'.split(',')}")
+    private List<String> validExtension;
+
+    @Value("${imageUploadUtil.max.upload-size:1}")
+    private Integer maxUploadSize;
+
     public String profileImage(String filename, MultipartFile file) {
+        if (!validExtension.contains(getExtension(file))) {
+            throw new ConnectDataException(
+                    ConnectErrorCode.PARAM_EXCEPTION,
+                    "invalid file extension found for profile image"
+            );
+        }
+        if (file.getSize() > maxUploadSize * 1024 * 1024) {
+            throw new ConnectDataException(
+                    ConnectErrorCode.PARAM_EXCEPTION,
+                    "exceed max file size for profile image (should below " + maxUploadSize + "MB"
+            );
+        }
+
         try {
             Path uploadPath = Paths.get(profileImageDir);
             if (!Files.exists(uploadPath)) {
                 Files.createDirectories(uploadPath);
             }
+
             Path filePath = uploadPath.resolve(filename);
             Files.deleteIfExists(filePath);
             Files.copy(file.getInputStream(), filePath);
+            return filePath.toString();
         } catch (IOException e) {
             throw new ConnectDataException(
                     ConnectErrorCode.INTERNAL_SERVER_ERROR,
                     "profile image upload failed :" + e.getMessage() + e.getCause()
             );
         }
-        return filename;
     }
 
     public String getExtension(MultipartFile file) {
